Added `DocumentAnalyzer` class that contains a lot of helper methods to analyze query. This includes newly introduced `deprecatedUsages` and `introspectionUsages` Closes #211 Closes #207
