+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Upcoming
 
+* Added `DocumentAnalyzer` class that contains a lot of helper methods to analyze query. This includes newly introduced `deprecatedUsages` and `introspectionUsages` (#211, #207) 
+* `ExecutionScheme.Extended` now returns updated user context for mutations (#209)
 * Added `Context.isIntrospection` and `sangria.introspection.isIntrospection` helper methods to easily distinguish between introspection types and fields in middleware and query reducers. 
 * Added `QueryReducer.hasIntrospection` and `QueryReducer.rejectIntrospection` that rejects queries which contain introspection fields (useful for production environments where introspection can potentially be abused). 
+* Added aliases for `graphql`/`graphqlInput` macros: `gql`/`gqlInp` 
 
 ## v1.0.0 (2017-01-16)
