+++ src/main/scala/sangria/macros/package.scala
@@ -12,8 +12,11 @@ import sangria.ast.{Document, Value}
 
 package object macros {
   implicit class LiteralGraphQLStringContext(val sc: StringContext) extends AnyVal {
-    def graphql(): Document = macro ParseMacro.impl
-    def graphqlInput(): Value = macro ParseMacro.implInput
+    def gql(args: Any*): Document = macro ParseMacro.impl
+    def gqlInp(args: Any*): Value = macro ParseMacro.implInput
+
+    def graphql(args: Any*): Document = macro ParseMacro.impl
+    def graphqlInput(args: Any*): Value = macro ParseMacro.implInput
   }
 
 }
