+++ src/main/scala/sangria/validation/QueryValidator.scala
@@ -1,15 +1,14 @@
 package sangria.validation
 
 import sangria.ast
-import sangria.ast.{AstVisitorCommand, AstVisitor, FragmentDefinition}
+import sangria.ast.{AstVisitor, AstVisitorCommand}
 import sangria.ast.AstVisitorCommand._
 import sangria.parser.SourceMapper
 import sangria.renderer.SchemaRenderer
 import sangria.schema._
-import sangria.introspection.{SchemaMetaField, TypeMetaField, TypeNameMetaField}
 import sangria.validation.rules._
-import scala.collection.concurrent.TrieMap
-import scala.collection.mutable.{Set ⇒ MutableSet, Map ⇒ MutableMap, ListBuffer}
+
+import scala.collection.mutable.{ListBuffer, Set ⇒ MutableSet, Map ⇒ MutableMap}
 
 trait QueryValidator {
   def validateQuery(schema: Schema[_, _], queryAst: ast.Document): Vector[Violation]
@@ -121,89 +120,6 @@ class ValidationContext(val schema: Schema[_, _], val doc: ast.Document, val typ
   def violations = errors.toVector
 }
 
-case class DocumentAnalyzer(schema: Schema[_, _], document: ast.Document) {
-  import DocumentAnalyzer._
-
-  private val fragmentSpreadsCache = TrieMap[Int, List[ast.FragmentSpread]]()
-  private val recursivelyReferencedFragmentsCache = TrieMap[Int, List[ast.FragmentDefinition]]()
-  private val variableUsages = TrieMap[Int, List[VariableUsage]]()
-  private val recursiveVariableUsages = TrieMap[Int, List[VariableUsage]]()
-
-  def getFragmentSpreads(astNode: ast.SelectionContainer) =
-    fragmentSpreadsCache.getOrElseUpdate(astNode.cacheKeyHash, {
-      val spreads = ListBuffer[ast.FragmentSpread]()
-      val setsToVisit = ValidatorStack.empty[List[ast.Selection]]
-
-      setsToVisit.push(astNode.selections)
-
-      while (setsToVisit.nonEmpty) {
-        val set = setsToVisit.pop()
-
-        set.foreach {
-          case fs: ast.FragmentSpread ⇒
-            spreads += fs
-          case cont: ast.SelectionContainer ⇒
-            setsToVisit push cont.selections
-        }
-      }
-
-      spreads.toList
-    })
-
-  def getRecursivelyReferencedFragments(operation: ast.OperationDefinition) =
-    recursivelyReferencedFragmentsCache.getOrElseUpdate(operation.cacheKeyHash, {
-      val frags = ListBuffer[ast.FragmentDefinition]()
-      val collectedNames = MutableSet[String]()
-      val nodesToVisit = ValidatorStack.empty[ast.SelectionContainer]
-
-      nodesToVisit.push(operation)
-
-      while (nodesToVisit.nonEmpty) {
-        val node = nodesToVisit.pop()
-        val spreads = getFragmentSpreads(node)
-
-        spreads.foreach { spread ⇒
-          val fragName = spread.name
-
-          if (!collectedNames.contains(fragName)) {
-            collectedNames += fragName
-
-            document.fragments.get(fragName) match {
-              case Some(frag) ⇒
-                frags += frag
-                nodesToVisit.push(frag)
-              case None ⇒ // do nothing
-            }
-          }
-        }
-      }
-
-      frags.toList
-    })
-
-  def getVariableUsages(astNode: ast.SelectionContainer) =
-    variableUsages.getOrElseUpdate(astNode.cacheKeyHash, {
-      AstVisitor.visitAstWithState(schema, astNode, ListBuffer[VariableUsage]()) { (typeInfo, usages) ⇒
-        AstVisitor {
-          case _: ast.VariableDefinition ⇒ Skip
-          case vv: ast.VariableValue ⇒
-            usages += VariableUsage(vv, typeInfo.inputType)
-            Continue
-        }
-      }.toList
-    })
-
-  def getRecursiveVariableUsages(operation: ast.OperationDefinition) =
-    recursiveVariableUsages.getOrElseUpdate(operation.cacheKeyHash,
-      getRecursivelyReferencedFragments(operation).foldLeft(getVariableUsages(operation)) {
-        case (acc, fragment) ⇒ acc ++ getVariableUsages(fragment)
-      })
-}
-
-object DocumentAnalyzer {
-  case class VariableUsage(node: ast.VariableValue, tpe: Option[InputType[_]])
-}
-
 object ValidationContext {
   def isValidLiteralValue(tpe: InputType[_], value: ast.Value, sourceMapper: Option[SourceMapper]): Vector[Violation] = (tpe, value) match {
     case (_, _: ast.VariableValue) ⇒ Vector.empty
@@ -250,170 +166,4 @@ object ValidationContext {
         case _ ⇒ Vector.empty
       }
   }
-}
-
-class TypeInfo(schema: Schema[_, _]) {
-  // Using mutable data-structures and mutability to minimize validation footprint
-
-  private val typeStack: ValidatorStack[Option[Type]] = ValidatorStack.empty
-  private val parentTypeStack: ValidatorStack[Option[CompositeType[_]]] = ValidatorStack.empty
-  private val inputTypeStack: ValidatorStack[Option[InputType[_]]] = ValidatorStack.empty
-  private val fieldDefStack: ValidatorStack[Option[Field[_, _]]] = ValidatorStack.empty
-  private val ancestorStack: ValidatorStack[ast.AstNode] = ValidatorStack.empty
-
-  var directive: Option[Directive] = None
-  var argument: Option[Argument[_]] = None
-
-  def tpe = typeStack.headOption.flatten
-  def previousParentType = parentTypeStack.headOption(1).flatten
-  def parentType = parentTypeStack.headOption.flatten
-  def inputType = inputTypeStack.headOption.flatten
-  def fieldDef = fieldDefStack.headOption.flatten
-  def ancestors: Seq[ast.AstNode] = ancestorStack.toSeq
-
-  def enter(node: ast.AstNode) = {
-    ancestorStack push node
-
-    node match {
-      case f: ast.Field ⇒
-        val parent = parentType
-        val fieldDef = parent flatMap (getFieldDef(_, f))
-        val fieldType = fieldDef map (_.fieldType)
-
-        fieldDefStack push fieldDef
-        typeStack push fieldType
-
-        pushParent()
-      case ast.Directive(name, _, _, _) ⇒
-        directive = schema.directivesByName get name
-      case ast.OperationDefinition(ast.OperationType.Query, _, _, _, _, _, _, _) ⇒
-        typeStack push Some(schema.query)
-        pushParent()
-      case ast.OperationDefinition(ast.OperationType.Mutation, _, _, _, _, _, _, _) ⇒
-        typeStack push schema.mutation
-        pushParent()
-      case ast.OperationDefinition(ast.OperationType.Subscription, _, _, _, _, _, _, _) ⇒
-        typeStack push schema.subscription
-        pushParent()
-      case fd: ast.FragmentDefinition ⇒
-        typeStack.push(schema.allTypes get fd.typeCondition.name)
-        pushParent()
-      case ifd: ast.InlineFragment ⇒
-        typeStack.push(ifd.typeCondition.fold(tpe)(schema.allTypes get _.name))
-        pushParent()
-      case vd: ast.VariableDefinition ⇒
-        inputTypeStack push schema.getInputType(vd.tpe)
-      case a: ast.Argument ⇒
-        argument = directive orElse fieldDef flatMap { withArgs ⇒
-          withArgs.arguments find (_.name == a.name)
-        }
-        inputTypeStack push argument.map(_.inputValueType)
-      case ast.ListValue(values, _, _) ⇒
-        inputType match {
-          case Some(it) ⇒ getNotNullType(it) match {
-            case it: ListInputType[_] ⇒ inputTypeStack push Some(it.ofType)
-            case _ ⇒ inputTypeStack push None
-          }
-          case None ⇒ inputTypeStack push None
-        }
-      case ast.ObjectField(name, value, _, _) ⇒
-        val fieldType = inputType flatMap (it ⇒ getNamedType(it) match {
-          case obj: InputObjectType[_] ⇒ obj.fieldsByName.get(name) map (_.inputValueType)
-          case _ ⇒ None
-        })
-
-        inputTypeStack push fieldType
-      case _ ⇒ // ignore
-    }
-  }
-
-  def pushParent(): Unit = {
-    tpe match {
-      case Some(some) ⇒ getNamedType(some) match {
-        case comp: CompositeType[_] ⇒ parentTypeStack push Some(comp)
-        case _ ⇒ parentTypeStack push None
-      }
-      case _ ⇒ parentTypeStack push None
-    }
-  }
-
-  def leave(node: ast.AstNode) = {
-    node match {
-      case f: ast.Field ⇒
-        fieldDefStack.pop()
-        typeStack.pop()
-        parentTypeStack.pop()
-      case ast.Directive(name, _, _, _) ⇒
-        directive = None
-      case ast.OperationDefinition(ast.OperationType.Query, _, _, _, _, _, _, _) ⇒
-        typeStack.pop()
-        parentTypeStack.pop()
-      case ast.OperationDefinition(ast.OperationType.Mutation, _, _, _, _, _, _, _) ⇒
-        typeStack.pop()
-        parentTypeStack.pop()
-      case ast.OperationDefinition(ast.OperationType.Subscription, _, _, _, _, _, _, _) ⇒
-        typeStack.pop()
-        parentTypeStack.pop()
-      case fd: ast.FragmentDefinition ⇒
-        typeStack.pop()
-        parentTypeStack.pop()
-      case fd: ast.InlineFragment ⇒
-        typeStack.pop()
-        parentTypeStack.pop()
-      case vd: ast.VariableDefinition ⇒
-        inputTypeStack.pop()
-      case a: ast.Argument ⇒
-        argument = None
-        inputTypeStack.pop()
-      case ast.ListValue(values, _, _) ⇒
-        inputTypeStack.pop()
-      case ast.ObjectField(name, value, _, _) ⇒
-        inputTypeStack.pop()
-      case _ ⇒ // ignore
-    }
-
-    ancestorStack.pop()
-  }
-
-  def getNamedType(it: Type): Type with Named = it match {
-    case OptionInputType(ofType) ⇒ getNamedType(ofType)
-    case OptionType(ofType) ⇒ getNamedType(ofType)
-    case ListInputType(ofType) ⇒ getNamedType(ofType)
-    case ListType(ofType) ⇒ getNamedType(ofType)
-    case n: Named ⇒ n
-    case t ⇒ throw new IllegalStateException("Expected named type, but got: " + t)
-  }
-
-  def getNotNullType(it: InputType[_]): InputType[_] = it match {
-    case OptionInputType(ofType) ⇒ ofType
-    case n ⇒ n
-  }
-
-  def getFieldDef(parent: CompositeType[_], astField: ast.Field): Option[Field[_, _]] = {
-    if (astField.name == SchemaMetaField.name && schema.query.name == parent.name)
-      Some(SchemaMetaField)
-    else if (astField.name == TypeMetaField.name && schema.query.name == parent.name)
-      Some(TypeMetaField)
-    else if (astField.name == TypeNameMetaField.name)
-      Some(TypeNameMetaField)
-    else parent match {
-      case o: ObjectLikeType[_, _] ⇒ o.getField(schema, astField.name).headOption
-      case _ ⇒ None
-    }
-  }
-}
-
-class ValidatorStack[T] {
-  private val stack: ListBuffer[T] = ListBuffer.empty
-    
-  def push(element: T): Unit = stack.prepend(element)
-  def pop(): T = stack.remove(0)
-  def headOption = stack.headOption
-  def headOption(toDrop: Int) = stack.drop(toDrop).headOption
-  def nonEmpty = stack.nonEmpty
-  def toSeq: Seq[T] = stack
-}
-
-object ValidatorStack {
-  def empty[T] = new ValidatorStack[T]
 }
